refactor(code-mode): isolate eager-only policy
